Broken down main into an application class and editor widgets
