+++ source/application/simulation/scene.cpp
@@ -1,22 +1,22 @@
 #include "scene.hpp"
 
-#include <file_utility.hpp>
+#include "utils/file_utility.hpp"
 
-namespace sim
+namespace app
 {
 	scene::scene(const star& disk_star, const char* orbits_file, const float fixed_radiation, const int particles_per_orbit, const char* filepath)
 		: filename{ filepath }
 	{
-		const std::vector<orbit>& orbits = load_orbits_from_file(orbits_file, fixed_radiation);
+		const std::vector<orbit>& orbits = utils::load_orbits_from_file(orbits_file, fixed_radiation);
 		particles = create_debris_disk(orbits, particles_per_orbit, disk_star);
 
-		save_particles_to_file(particles, filepath);
+		utils::save_particles_to_file(particles, filepath);
 	}
 
 	scene::scene(const char* filepath)
 		: filename{ filepath }
 	{
-		particles = load_particles_from_file(filename);
+		particles = utils::load_particles_from_file(filename);
 	}
 
 	const std::vector<particle>& scene::get_particles() const
