--- source/engine.cpp
@@ -1,331 +1,0 @@
-#include "engine.hpp"
-
-#include "log.hpp"
-
-#include <glad/glad.h>
-#include <GLFW/glfw3.h>
-
-#define STB_IMAGE_IMPLEMENTATION
-#include "stb_image/stb_image.h"
-
-#include <fstream>
-#include <sstream>
-
-namespace
-{
-	void on_window_resize(GLFWwindow* window, const int width, const int height)
-	{
-		glViewport(0, 0, width, height);
-	}
-
-#ifdef DEBUG
-	void check_shader_compile_errors(const unsigned int shaderID, const char* type)
-	{
-		int success;
-		char infoLog[1024];
-
-		if (type != "PROGRAM")
-		{
-			glGetShaderiv(shaderID, GL_COMPILE_STATUS, &success);
-			glGetShaderInfoLog(shaderID, 1024, NULL, infoLog);
-
-			ASSERT(success, return, "SHADER_COMPILATION_ERROR of type: {0}\n{1}", type, infoLog)
-		}
-		else
-		{
-			glGetProgramiv(shaderID, GL_LINK_STATUS, &success);
-			glGetProgramInfoLog(shaderID, 1024, NULL, infoLog);
-
-			ASSERT(success, return, "PROGRAM_LINKING_ERROR of type: {0}\n{1}", type, infoLog)
-		}
-	}
-#endif
-}
-
-namespace sim
-{
-	window::window(const int width, const int height, const char* title)
-		: native_window{ nullptr }
-	{
-		glfwInit();
-		glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 4);
-		glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 5);
-		glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-
-		native_window = glfwCreateWindow(width, height, title, nullptr, nullptr);
-
-		glfwMakeContextCurrent(native_window);
-		glfwSetFramebufferSizeCallback(native_window, on_window_resize);
-
-		gladLoadGLLoader((GLADloadproc)glfwGetProcAddress);
-	}
-
-	void window::update()
-	{
-		glfwPollEvents();
-		glfwSwapBuffers(native_window);
-	}
-
-	bool window::is_open() const
-	{
-		return !glfwWindowShouldClose(native_window);
-	}
-
-	window::~window()
-	{
-		glfwDestroyWindow(native_window);
-		glfwTerminate();
-	}
-
-	shader::shader(const char* vertex_path, const char* fragment_path)
-	{
-		std::ifstream v_shader_file;
-		std::ifstream f_shader_file;
-
-		v_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-		f_shader_file.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-
-		v_shader_file.open(vertex_path);
-		f_shader_file.open(fragment_path);
-		std::stringstream v_shader_stream, f_shader_stream;
-
-		v_shader_stream << v_shader_file.rdbuf();
-		f_shader_stream << f_shader_file.rdbuf();
-
-		v_shader_file.close();
-		f_shader_file.close();
-
-		std::string vertex_code = v_shader_stream.str();
-		std::string fragment_code = f_shader_stream.str();
-		const char* v_shader_code = vertex_code.c_str();
-		const char* f_shader_code = fragment_code.c_str();
-
-		unsigned int vertex = glCreateShader(GL_VERTEX_SHADER);
-		glShaderSource(vertex, 1, &v_shader_code, nullptr);
-		glCompileShader(vertex);
-
-#ifdef DEBUG
-		check_shader_compile_errors(vertex, "VERTEX");
-#endif
-
-		unsigned int fragment = glCreateShader(GL_FRAGMENT_SHADER);
-		glShaderSource(fragment, 1, &f_shader_code, nullptr);
-		glCompileShader(fragment);
-
-#ifdef DEBUG
-		check_shader_compile_errors(fragment, "FRAGMENT");
-#endif
-
-		id = glCreateProgram();
-		glAttachShader(id, vertex);
-		glAttachShader(id, fragment);
-		glLinkProgram(id);
-
-#ifdef DEBUG
-		check_shader_compile_errors(id, "PROGRAM");
-#endif
-
-		glDeleteShader(vertex);
-		glDeleteShader(fragment);
-	}
-
-	shader::~shader()
-	{
-		glDeleteProgram(id);
-	}
-
-	void shader::use() const
-	{
-		glUseProgram(id);
-	}
-
-	template<>
-	void shader::set_uniform<bool>(const char* name, const bool& value) const
-	{
-		glUniform1i(glGetUniformLocation(id, name), (int)value);
-	}
-
-	template<>
-	void shader::set_uniform<int>(const char* name, const int& value) const
-	{
-		glUniform1i(glGetUniformLocation(id, name), value);
-	}
-
-	template<>
-	void shader::set_uniform<float>(const char* name, const float& value) const
-	{
-		glUniform1f(glGetUniformLocation(id, name), value);
-	}
-
-	template<>
-	void shader::set_uniform<glm::vec2>(const char* name, const glm::vec2& value) const
-	{
-		glUniform2fv(glGetUniformLocation(id, name), 1, &value[0]);
-	}
-
-	template<>
-	void shader::set_uniform<glm::vec3>(const char* name, const glm::vec3& value) const
-	{
-		glUniform3fv(glGetUniformLocation(id, name), 1, &value[0]);
-	}
-
-	template<>
-	void shader::set_uniform<glm::vec4>(const char* name, const glm::vec4& value) const
-	{
-		glUniform4fv(glGetUniformLocation(id, name), 1, &value[0]);
-	}
-
-	template<>
-	void shader::set_uniform<glm::mat4>(const char* name, const glm::mat4& value) const
-	{
-		glUniformMatrix4fv(glGetUniformLocation(id, name), 1, GL_FALSE, &value[0][0]);
-	}
-
-	template<>
-	void shader::set_uniform<texture_1d>(const char* name, const texture_1d& value) const
-	{
-		glUniform1i(glGetUniformLocation(id, name), value.get_unit());
-	}
-
-	texture_1d::texture_1d(const char* filename, const unsigned int unit)
-		: unit{ unit }
-	{
-		glGenTextures(1, &id);
-		glBindTexture(GL_TEXTURE_1D, id);
-
-		glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-		glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-		glTexParameteri(GL_TEXTURE_1D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-
-		int width, height, channel_count;
-		unsigned char* texture_data = stbi_load(filename, &width, &height, &channel_count, 0);
-
-		ASSERT(texture_data != nullptr, return, "Failed to load texture %s", filename);
-
-		glTexImage1D(GL_TEXTURE_1D, 0, GL_RGB, width, 0, GL_RGB, GL_UNSIGNED_BYTE, texture_data);
-		stbi_image_free(texture_data);
-
-		glBindTexture(GL_TEXTURE_1D, 0);
-	}
-
-	texture_1d::~texture_1d()
-	{
-		glDeleteTextures(1, &id);
-	}
-
-	void texture_1d::bind() const
-	{
-		glActiveTexture(GL_TEXTURE0 + unit);
-		glBindTexture(GL_TEXTURE_1D, id);
-	}
-
-	void texture_1d::unbind() const
-	{
-		glActiveTexture(GL_TEXTURE0 + unit);
-		glBindTexture(GL_TEXTURE_1D, 0);
-	}
-
-	unsigned int texture_1d::get_unit() const
-	{
-		return unit;
-	}
-
-	renderer::renderer()
-		: particle_count{ 0 }
-		, particle_buffer{ 0 }
-		, framebuffer{ 0 }
-		, render_texture{ 0 }
-	{
-		glGenVertexArrays(1, &vao);
-
-		glEnable(GL_PROGRAM_POINT_SIZE);
-		glEnable(GL_BLEND);
-		glBlendFunc(GL_SRC_ALPHA, GL_ONE);
-		glDisable(GL_DEPTH_TEST);
-	}
-
-	renderer::~renderer()
-	{
-		glDeleteVertexArrays(1, &vao);
-
-		if (framebuffer != 0)
-			glDeleteFramebuffers(1, &framebuffer);
-
-		if (render_texture != 0)
-			glDeleteTextures(1, &render_texture);
-
-		if (particle_buffer != 0)
-			glDeleteBuffers(1, &particle_buffer);
-	}
-
-	void renderer::load_particles(const size_t particle_count, const size_t particle_size, const void* particles)
-	{
-		ASSERT(particles != nullptr, return, "particles were nullptr");
-
-		if (particle_buffer != 0)
-			glDeleteBuffers(1, &particle_buffer);
-
-		this->particle_count = particle_count;
-
-		glGenBuffers(1, &particle_buffer);
-
-		glBindBuffer(GL_SHADER_STORAGE_BUFFER, particle_buffer);
-		glBufferData(GL_SHADER_STORAGE_BUFFER, particle_count * particle_size, particles, GL_STATIC_DRAW);
-		glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-	}
-
-	void renderer::create_framebuffer(const int width, const int height)
-	{
-		if (framebuffer != 0)
-			glDeleteFramebuffers(1, &framebuffer);
-
-		if (render_texture != 0)
-			glDeleteTextures(1, &render_texture);
-
-		glGenFramebuffers(1, &framebuffer);
-		glBindFramebuffer(GL_FRAMEBUFFER, framebuffer);
-
-		glGenTextures(1, &render_texture);
-		glBindTexture(GL_TEXTURE_2D, render_texture);
-
-		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA, GL_UNSIGNED_BYTE, nullptr);
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-		glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, render_texture, 0);
-	}
-
-	void renderer::bind_framebuffer(const int width, const int height)
-	{
-		glBindFramebuffer(GL_FRAMEBUFFER, framebuffer);
-		glViewport(0, 0, width, height);
-	}
-
-	void renderer::unbind_framebuffer()
-	{
-		glBindFramebuffer(GL_FRAMEBUFFER, 0);
-	}
-
-	std::vector<unsigned char> renderer::read_pixels(const int width, const int height) const
-	{
-		std::vector<unsigned char> pixels(width * height * 3);
-		glReadPixels(0, 0, width, height, GL_RGB, GL_UNSIGNED_BYTE, pixels.data());
-		return pixels;
-	}
-
-	void renderer::clear()
-	{
-		glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
-		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-	}
-
-	void renderer::render()
-	{
-		if (particle_count == 0)
-			return;
-
-		glBindVertexArray(vao);
-		glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 0, particle_buffer);
-
-		glDrawArraysInstanced(GL_POINTS, 0, 1, static_cast<GLsizei>(particle_count));
-	}
-}
