+++ source/engine/editor/widget.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace engine
+{
+    class widget
+    {
+    public:
+        virtual ~widget() = default;
+        virtual void draw() = 0;
+    };
+}
